added fullscreen capability, key callback for glfw window and fixed floating point warnings

+++ V/renderer/Camera/Camera.cc
@@ -28,60 +28,59 @@ void v::renderer::Camera::Matrix(Shader & shader, const char * uniform) {
 }
 
 void v::renderer::Camera::Inputs(GLFWwindow * window) {
-    if (glfwGetKey(window, GLFW_KEY_W) == GLFW_PRESS) 
-		Position += speed * Orientation;
+    if(glfwGetKey(window, GLFW_KEY_W) == GLFW_PRESS) 
+        Position += speed * Orientation;
 	
-	if (glfwGetKey(window, GLFW_KEY_A) == GLFW_PRESS) 
-		Position += speed * -glm::normalize(glm::cross(Orientation, Up));
+    if(glfwGetKey(window, GLFW_KEY_A) == GLFW_PRESS) 
+	Position += speed * -glm::normalize(glm::cross(Orientation, Up));
 	
-	if (glfwGetKey(window, GLFW_KEY_S) == GLFW_PRESS) 
-		Position += speed * -Orientation;
+    if(glfwGetKey(window, GLFW_KEY_S) == GLFW_PRESS) 
+	Position += speed * -Orientation;
 	
-	if (glfwGetKey(window, GLFW_KEY_D) == GLFW_PRESS) 
-		Position += speed * glm::normalize(glm::cross(Orientation, Up));
+    if(glfwGetKey(window, GLFW_KEY_D) == GLFW_PRESS) 
+        Position += speed * glm::normalize(glm::cross(Orientation, Up));
 	
-	if (glfwGetKey(window, GLFW_KEY_SPACE) == GLFW_PRESS) 
-		Position += speed * Up;
+    if(glfwGetKey(window, GLFW_KEY_SPACE) == GLFW_PRESS) 
+	Position += speed * Up;
 	
-	if (glfwGetKey(window, GLFW_KEY_LEFT_CONTROL) == GLFW_PRESS) 
-		Position += speed * -Up;
+    if(glfwGetKey(window, GLFW_KEY_LEFT_CONTROL) == GLFW_PRESS) 
+	Position += speed * -Up;
 	
-	if (glfwGetKey(window, GLFW_KEY_LEFT_SHIFT) == GLFW_PRESS) 
-		speed = 0.4f;
+    if(glfwGetKey(window, GLFW_KEY_LEFT_SHIFT) == GLFW_PRESS) 
+	speed = 0.4f;
 	
-	else if (glfwGetKey(window, GLFW_KEY_LEFT_SHIFT) == GLFW_RELEASE) 
-		speed = 0.1f;
+    else if(glfwGetKey(window, GLFW_KEY_LEFT_SHIFT) == GLFW_RELEASE) 
+	speed = 0.1f;
 	//  \/ you will need to remove this if and comment out firstClick for the camera to work all the time
-	if (glfwGetMouseButton(window, GLFW_MOUSE_BUTTON_LEFT) == GLFW_PRESS) {
-		glfwSetInputMode(window, GLFW_CURSOR, GLFW_CURSOR_HIDDEN);
+    if(glfwGetMouseButton(window, GLFW_MOUSE_BUTTON_LEFT) == GLFW_PRESS) {
+	glfwSetInputMode(window, GLFW_CURSOR, GLFW_CURSOR_HIDDEN);
 
-		if (firstClick) {
-			glfwSetCursorPos(window, (width / 2), (height / 2));
-			firstClick = false;
-		}
+	if (firstClick) {
+    	    glfwSetCursorPos(window, ((double)width / 2.0), ((double)height / 2.0));
+	    firstClick = false;
+	}
 
-		double mouseX;
-		double mouseY;
+	double mouseX;
+        double mouseY;
 
-		glfwGetCursorPos(window, &mouseX, &mouseY);
+        glfwGetCursorPos(window, &mouseX, &mouseY);
 
-		float rotX = sensitivity * (float)(mouseY - (height / 2)) / height;
-		float rotY = sensitivity * (float)(mouseX - (width / 2)) / width;
+	float rotX = sensitivity * (float)(mouseY - ((float)height / 2.0F)) / (float)height;
+	float rotY = sensitivity * (float)(mouseX - ((float)width / 2.0F)) / (float)width;
 
-		glm::vec3 newOrientation = glm::rotate(Orientation, glm::radians(-rotX), glm::normalize(glm::cross(Orientation, Up)));
+	glm::vec3 newOrientation = glm::rotate(Orientation, glm::radians(-rotX), glm::normalize(glm::cross(Orientation, Up)));
 
-		if (abs(glm::angle(newOrientation, Up) - glm::radians(90.0f)) <= glm::radians(85.0f))
-			Orientation = newOrientation;
+	if(abs(glm::angle(newOrientation, Up) - glm::radians(90.0f)) <= glm::radians(85.0f))
+	    Orientation = newOrientation;
 		
 
-		Orientation = glm::rotate(Orientation, glm::radians(-rotY), Up);
+	Orientation = glm::rotate(Orientation, glm::radians(-rotY), Up);
 
-		glfwSetCursorPos(window, (width / 2), (height / 2));
+	glfwSetCursorPos(window, ((double)width / 2.0), ((double)height / 2.0));
 
-	} else if (glfwGetMouseButton(window, GLFW_MOUSE_BUTTON_LEFT) == GLFW_RELEASE) {
-
-		glfwSetInputMode(window, GLFW_CURSOR, GLFW_CURSOR_NORMAL);
-
-		firstClick = true;
-	}
+    } else if(glfwGetMouseButton(window, GLFW_MOUSE_BUTTON_LEFT) == GLFW_RELEASE) {
+        glfwSetInputMode(window, GLFW_CURSOR, GLFW_CURSOR_NORMAL);
+            
+        firstClick = true;
+    }
 }
